fix(modules): align .h declarations to autolink *Spec (systemic)
The autolink *Spec is the contract; the legacy .h method declarations
(using NSNumber* params etc.) must match. align_spec.py now also rewrites
.h declarations. Fixes crypto's getRandomBytes:(NSNumber*) vs spec (double).

NOTE: apple-authentication is a delegate/callback-based async module that
autolink v0.4.1 synchronous-return cannot express — excluded from green
pending an async-bridge decision.

+++ packages/modules/brightness/ios/src/BrightnessModule.h
@@ -17,7 +17,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)setBrightness:(double)brightnessValue;
 - (double)getSystemBrightness;
 - (BOOL)isUsingSystemBrightness;
-- (int)getSystemBrightnessMode;
+- (double)getSystemBrightnessMode;
 @end
 
 NS_ASSUME_NONNULL_END
